sim_bridge/requirements.txt
# PX4 SITL Bridge Requirements

# MAVLink and PX4 communication
mavsdk>=1.4.0
aiomavlink>=0.1.0

# WebSocket client
websockets>=12.0

# Configuration
PyYAML>=6.0

# Coordinate transformations
pyproj>=3.6.0

# Async utilities
asyncio-mqtt>=0.16.1

# Logging
colorlog>=6.7.0
